--- Dockerfile
@@ -1,11 +1,0 @@
-# Fichier Dockerfile pour le service REST
-FROM python:3.8-slim
-
-WORKDIR /app
-
-COPY requirements.txt requirements.txt
-RUN pip install --no-cache-dir -r requirements.txt
-
-COPY . .
-
-CMD ["python", "app.py"]
