+++ apache-config/rest_service.conf
@@ -1,0 +1,10 @@
+<VirtualHost *:80>
+    ServerName rest.example.com
+
+    ProxyPreserveHost On
+    ProxyPass / http://rest:5000/
+    ProxyPassReverse / http://rest:5000/
+
+    ErrorLog ${APACHE_LOG_DIR}/rest_error.log
+    CustomLog ${APACHE_LOG_DIR}/rest_access.log combined
+</VirtualHost>
