+++ apache-config/rpc_service.conf
@@ -1,0 +1,10 @@
+<VirtualHost *:80>
+    ServerName rpc.example.com
+
+    ProxyPreserveHost On
+    ProxyPass / http://rpc:9000/
+    ProxyPassReverse / http://rpc:9000/
+
+    ErrorLog ${APACHE_LOG_DIR}/rpc_error.log
+    CustomLog ${APACHE_LOG_DIR}/rpc_access.log combined
+</VirtualHost>
