+++ apache-config/soap_service.conf
@@ -1,0 +1,10 @@
+<VirtualHost *:80>
+    ServerName soap.example.com
+
+    ProxyPreserveHost On
+    ProxyPass / http://soap:8000/
+    ProxyPassReverse / http://soap:8000/
+
+    ErrorLog ${APACHE_LOG_DIR}/soap_error.log
+    CustomLog ${APACHE_LOG_DIR}/soap_access.log combined
+</VirtualHost>
