+++ rest_service/Dockerfile
@@ -1,0 +1,18 @@
+# Utiliser une image Python de base
+FROM python:3.9-slim
+
+# Définir le répertoire de travail
+WORKDIR /app
+
+# Copier les fichiers requirements.txt et installer les dépendances
+COPY requirements.txt .
+RUN pip install --no-cache-dir -r requirements.txt
+
+# Copier le reste du code de l'application
+COPY . .
+
+# Exposer le port sur lequel l'application va s'exécuter
+EXPOSE 7777
+
+# Définir la commande pour exécuter l'application
+CMD ["python", "app.py"]
