+++ rest_service/requirements.txt
@@ -1,0 +1,3 @@
+flask
+flask-restful
+flask-swagger-ui
