--- soap_service.py
@@ -1,27 +1,0 @@
-# Fichier soap_service.py
-from spyne import Application, rpc, ServiceBase, Iterable, Integer, Unicode
-from spyne.protocol.soap import Soap11
-from spyne.server.wsgi import WsgiApplication
-
-class HelloWorldService(ServiceBase):
-    @rpc(Unicode, Integer, _returns=Iterable(Unicode))
-    def say_hello(ctx, name, times):
-        for i in range(times):
-            yield 'Hello, %s' % name
-
-application = Application([HelloWorldService],
-    tns='namespace',
-    in_protocol=Soap11(validator='lxml'),
-    out_protocol=Soap11())
-
-wsgi_application = WsgiApplication(application)
-
-if __name__ == '__main__':
-    import logging
-
-    from wsgiref.simple_server import make_server
-    logging.basicConfig(level=logging.DEBUG)
-    logging.getLogger('spyne.protocol.xml').setLevel(logging.DEBUG)
-
-    server = make_server('0.0.0.0', 8000, wsgi_application)
-    server.serve_forever()
