+++ soap_service/requirements.txt
@@ -1,0 +1,2 @@
+spyne
+lxml
